fixed integation test race condition
