+++ src/lib.rs
@@ -865,126 +865,11 @@ mod tests {
         assert!(!exists_error.is_not_found());
     }
 
-    #[test]
-    fn test_flashq_with_file_storage() {
-        use crate::storage::{StorageBackend, file::SyncMode};
-        use std::env;
+    // File-based test moved to tests/storage_integration_tests.rs
 
-        let temp_dir = std::env::temp_dir().join(format!("flashq_integration_test_{}", std::process::id()));
-        let original_dir = env::current_dir().unwrap();
-        
-        std::fs::create_dir_all(&temp_dir).unwrap();
-        env::set_current_dir(&temp_dir).unwrap();
+    // File-based test moved to tests/storage_integration_tests.rs
 
-        let queue = FlashQ::with_storage_backend(StorageBackend::File(SyncMode::Immediate));
-
-        // Test basic operations
-        let record1 = Record::new(Some("key1".to_string()), "value1".to_string(), None);
-        let offset1 = queue.post_record("test_topic".to_string(), record1).unwrap();
-        assert_eq!(offset1, 0);
-
-        let record2 = Record::new(Some("key2".to_string()), "value2".to_string(), None);  
-        let offset2 = queue.post_record("test_topic".to_string(), record2).unwrap();
-        assert_eq!(offset2, 1);
-
-        // Test polling
-        let records = queue.poll_records("test_topic", None).unwrap();
-        assert_eq!(records.len(), 2);
-        assert_eq!(records[0].record.value, "value1");
-        assert_eq!(records[1].record.value, "value2");
-
-        // Test different topic
-        let record3 = Record::new(None, "value3".to_string(), None);
-        let offset3 = queue.post_record("other_topic".to_string(), record3).unwrap();
-        assert_eq!(offset3, 0); // New topic starts at offset 0
-
-        let records_other = queue.poll_records("other_topic", None).unwrap();
-        assert_eq!(records_other.len(), 1);
-        assert_eq!(records_other[0].record.value, "value3");
-
-        // Test that original topic still has its records
-        let records_original = queue.poll_records("test_topic", None).unwrap();
-        assert_eq!(records_original.len(), 2);
-
-        env::set_current_dir(original_dir).unwrap();
-        std::fs::remove_dir_all(&temp_dir).ok();
-    }
-
-    #[test]
-    fn test_file_storage_vs_memory_storage_compatibility() {
-        use crate::storage::{StorageBackend, file::SyncMode};
-
-        // Test that both backends provide the same interface and behavior
-        let memory_queue = FlashQ::with_storage_backend(StorageBackend::Memory);
-        let record = Record::new(Some("test_key".to_string()), "test_value".to_string(), None);
-        
-        let memory_offset = memory_queue.post_record("compatibility_test".to_string(), record.clone()).unwrap();
-        let memory_records = memory_queue.poll_records("compatibility_test", None).unwrap();
-
-        // File storage should behave identically
-        use std::env;
-        let temp_dir = std::env::temp_dir().join(format!("flashq_compatibility_test_{}", std::process::id()));
-        let original_dir = env::current_dir().unwrap();
-        
-        std::fs::create_dir_all(&temp_dir).unwrap();
-        env::set_current_dir(&temp_dir).unwrap();
-
-        let file_queue = FlashQ::with_storage_backend(StorageBackend::File(SyncMode::Immediate));
-        let file_offset = file_queue.post_record("compatibility_test".to_string(), record).unwrap();
-        let file_records = file_queue.poll_records("compatibility_test", None).unwrap();
-
-        // Both should give same results
-        assert_eq!(memory_offset, file_offset);
-        assert_eq!(memory_records.len(), file_records.len());
-        assert_eq!(memory_records[0].record.value, file_records[0].record.value);
-        assert_eq!(memory_records[0].record.key, file_records[0].record.key);
-
-        env::set_current_dir(original_dir).unwrap();
-        std::fs::remove_dir_all(&temp_dir).ok();
-    }
-
-    #[test]
-    fn test_file_storage_persistence() {
-        use crate::storage::{StorageBackend, file::SyncMode};
-        use std::env;
-
-        let temp_dir = std::env::temp_dir().join(format!("flashq_persistence_test_{}", std::process::id()));
-        let original_dir = env::current_dir().unwrap();
-        
-        std::fs::create_dir_all(&temp_dir).unwrap();
-        env::set_current_dir(&temp_dir).unwrap();
-
-        // Create queue and add records
-        {
-            let queue = FlashQ::with_storage_backend(StorageBackend::File(SyncMode::Immediate));
-            let record1 = Record::new(None, "persistent_value1".to_string(), None);
-            let record2 = Record::new(None, "persistent_value2".to_string(), None);
-            
-            queue.post_record("persistent_topic".to_string(), record1).unwrap();
-            queue.post_record("persistent_topic".to_string(), record2).unwrap();
-        } // Queue goes out of scope
-
-        // Create new queue instance - should have persistent data
-        {
-            let queue = FlashQ::with_storage_backend(StorageBackend::File(SyncMode::Immediate));
-            
-            // Post a new record to trigger topic creation and recovery
-            let record3 = Record::new(None, "new_value".to_string(), None);
-            let offset = queue.post_record("persistent_topic".to_string(), record3).unwrap();
-            
-            // Should start from offset 2 (after the recovered records)
-            assert_eq!(offset, 2);
-            
-            let records = queue.poll_records("persistent_topic", None).unwrap();
-            assert_eq!(records.len(), 3);
-            assert_eq!(records[0].record.value, "persistent_value1");
-            assert_eq!(records[1].record.value, "persistent_value2");
-            assert_eq!(records[2].record.value, "new_value");
-        }
-
-        env::set_current_dir(original_dir).unwrap();
-        std::fs::remove_dir_all(&temp_dir).ok();
-    }
+    // File-based test moved to tests/storage_integration_tests.rs
 }
 // =============================================================================
 // RECORD TESTS
